requirements.txt
dj-database-url==0.5.0
Django==2.2.13
gunicorn==20.0.4
psycopg2==2.8.5
python-dotenv==0.13.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
